Move all portaling logic into `BottomSheet`, remove from our own `Dialog` wrapper (#5661)
