+++ modules/bottom-sheet/src/lib/Portal.tsx
@@ -12,9 +12,7 @@ type ComponentMap = {
   [id: string]: Component
 }
 
-export type PortalComponent = ({children}: {children?: React.ReactNode}) => null
-
-export function createPortalGroup() {
+export function createPortalGroup_INTERNAL() {
   const Context = React.createContext<ContextType>({
     outlet: null,
     append: () => {},
@@ -67,8 +65,3 @@ export function createPortalGroup() {
 
   return {Provider, Outlet, Portal}
 }
-
-const DefaultPortal = createPortalGroup()
-export const Provider = DefaultPortal.Provider
-export const Outlet = DefaultPortal.Outlet
-export const Portal = DefaultPortal.Portal
